my-flask-app/requirements.txt
Flask==2.0.3
SQLAlchemy==1.4.22
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.3.1
Flask-Cors==3.10.9
Pillow==8.4.0
requests==2.26.0
opencv-python==4.5.3.20210927
pytest==6.2.5
python-dotenv==0.19.2